# backend/docs/swagger.json
{"swagger":"2.0","info":{"title":"QR Code Generator API","contact":{}},"basePath":"/api/v1","paths":{"/links":{"get":{"description":"Get all links created by the authenticated user","produces":["application/json"],"tags":["links"],"summary":"Get all links for a user","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/dto.GetAllLinksResponse"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/links/create":{"post":{"description":"Create a new shortened link for the authenticated user","consumes":["application/json"],"produces":["application/json"],"tags":["links"],"summary":"Create a new link","parameters":[{"description":"Link data","name":"link","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.CreateLinkRequest"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/dto.CreateLinkResponse"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/links/{id}":{"get":{"description":"Get a specific link by its ID for the authenticated user","produces":["application/json"],"tags":["links"],"summary":"Get a link by ID","parameters":[{"type":"integer","description":"Link ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/dto.GetLinkResponse"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}},"patch":{"description":"Edit a specific link by its ID for the authenticated user","consumes":["application/json"],"produces":["application/json"],"tags":["links"],"summary":"Edit a link","parameters":[{"type":"integer","description":"Link ID","name":"id","in":"path","required":true},{"description":"Updated link data","name":"link","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.EditLinkRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/dto.EditLinkResponse"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/links/{id}/download":{"get":{"description":"Download a QR code for a specific link by its ID in various formats (png, svg, pdf)","produces":["application/octet-stream"],"tags":["links"],"summary":"Download a QR code for a link","parameters":[{"type":"integer","description":"Link ID","name":"id","in":"path","required":true},{"enum":["png","svg","pdf"],"type":"string","description":"File type (png, svg, pdf)","name":"type","in":"query"}],"responses":{"200":{"description":"Returns the QR code file for download","schema":{"type":"string"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/links/{id}/transitions":{"get":{"description":"Get transition analytics for a specific link by its ID","produces":["application/json"],"tags":["links"],"summary":"Get transitions for a link","parameters":[{"type":"integer","description":"Link ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/dto.GetTransitionsResponse"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/login":{"post":{"description":"Log in a user with email and password, returns a JWT token in a cookie","consumes":["application/json"],"produces":["application/json"],"tags":["auth"],"summary":"Log in a user","parameters":[{"description":"User login data","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/qrcode":{"post":{"description":"Generate a QR code for a given URL with custom styling","consumes":["application/json"],"produces":["image/png"],"tags":["qrcode"],"summary":"Generate a QR code","parameters":[{"description":"QR code generation data","name":"qrcode","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.GenerateQRCodeRequest"}}],"responses":{"201":{"description":"Returns the generated QR code as a PNG image","schema":{"type":"string"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/redirect/{hash}":{"get":{"description":"Redirects a shortened link to its original URL","tags":["redirect"],"summary":"Redirect to original URL","parameters":[{"type":"string","description":"Link hash","name":"hash","in":"path","required":true}],"responses":{"302":{"description":"Redirects to the original URL","schema":{"type":"string"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}},"/register":{"post":{"description":"Register a new user with name, email, and password","consumes":["application/json"],"produces":["application/json"],"tags":["auth"],"summary":"Register a new user","parameters":[{"description":"User registration data","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.RegisterRequest"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/dto.RegisterResponse"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/dto.GenericError"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/dto.GenericError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/dto.GenericError"}}}}}},"definitions":{"dto.CreateLinkRequest":{"type":"object","required":["original_url"],"properties":{"original_url":{"type":"string"}}},"dto.CreateLinkResponse":{"type":"object","properties":{"id":{"type":"integer"},"message":{"type":"string"}}},"dto.EditLinkRequest":{"type":"object","required":["background","color","original_url"],"properties":{"background":{"type":"string"},"color":{"type":"string"},"original_url":{"type":"string"},"smoothing":{"type":"number","maximum":0.5,"minimum":0}}},"dto.EditLinkResponse":{"type":"object","properties":{"id":{"type":"integer"},"message":{"type":"string"}}},"dto.GenerateQRCodeRequest":{"type":"object","required":["background","color","url"],"properties":{"background":{"type":"string"},"color":{"type":"string"},"smoothing":{"type":"number","maximum":0.5,"minimum":0},"url":{"type":"string"}}},"dto.GenericError":{"type":"object","properties":{"error":{"type":"string","example":"Some error message"}}},"dto.GetAllLinksResponse":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/definitions/dto.LinkInfo"}},"message":{"type":"string"}}},"dto.GetLinkResponse":{"type":"object","properties":{"background":{"type":"string"},"color":{"type":"string"},"created_at":{"type":"string"},"hash":{"type":"string"},"id":{"type":"integer"},"original_url":{"type":"string"},"smoothing":{"type":"number"},"updated_at":{"type":"string"}}},"dto.GetTransitionsResponse":{"type":"object","properties":{"transitions":{"type":"array","items":{"$ref":"#/definitions/dto.TransitionItem"}}}},"dto.LinkInfo":{"type":"object","properties":{"id":{"type":"integer"},"original_url":{"type":"string"}}},"dto.LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"dto.RegisterRequest":{"type":"object","required":["email","name","password","second_password"],"properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string","minLength":8},"second_password":{"type":"string"}}},"dto.RegisterResponse":{"type":"object","properties":{"message":{"type":"string"}}},"dto.TransitionItem":{"type":"object","properties":{"browser":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"os":{"type":"string"},"referer":{"type":"string"},"user_agent":{"type":"string"}}}}}